Move OutsideLifecycleException to new artifact
