+++ lifecycle/autodispose-lifecycle/src/main/java/com/uber/autodispose/lifecycle/LifecycleNotStartedException.java
@@ -16,8 +16,6 @@
 
 package com.uber.autodispose.lifecycle;
 
-import com.uber.autodispose.OutsideLifecycleException;
-
 /**
  * Signifies an error occurred due to execution starting before the lifecycle has started.
  */
